KANGKAI CONTINUE

+++ libs/hbb_common/src/config.rs
@@ -65,7 +65,7 @@ const CHARS: &'static [char] = &[
 ];
 
 pub const KANGKAI_ID_SERVER_URL: &'static str = "http://localhost:8000/desk_code";
-pub const KANGKAI_ID_REFRESH_TIMEOUT: u64 = 3;
+pub const KANGKAI_ID_REFRESH_TIMEOUT: f32 = 3.0;
 
 pub const RENDEZVOUS_SERVERS: &'static [&'static str] = &[
     "rs-ny.rustdesk.com",
@@ -95,7 +95,7 @@ pub struct Config {
     #[serde(default)]
     kangkai_id_server_url: String,
     #[serde(default)]
-    kangkai_id_refresh_timeout: u64,
+    kangkai_id_refresh_timeout: f32,
     #[serde(default)]
     salt: String,
     #[serde(default)]
@@ -730,9 +730,9 @@ impl Config {
         CONFIG.read().unwrap().password.clone()
     }
 
-    pub fn get_kangkai_id_refresh_timeout() -> u64 {
+    pub fn get_kangkai_id_refresh_timeout() -> f32 {
         let server = CONFIG.read().unwrap().kangkai_id_refresh_timeout;
-        if server < 1 {
+        if server < 1.0 {
             KANGKAI_ID_REFRESH_TIMEOUT
         } else {
             server
